patch from Harrie to fix variable types (int -> size_t)
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V4-2-patches@5076 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/snmp_api.c
@@ -3942,8 +3942,8 @@ _sess_read(void *sessp,
     struct snmp_pdu *pdu;
     struct request_list *rp, *orp = NULL;
     int ret;
-    int addrlen;
-    int fromlength;
+    size_t addrlen;
+    size_t fromlength;
 
     sp = slp->session; isp = slp->internal;
     if (!sp || !isp) {
